+++ CHANGELOG.rst
@@ -122,6 +122,7 @@ Version 2022-dev
 -  added documentation about architecture to votca and votca-xtp (#869)
 -  check for gmx-2019 in csg-tutorials (#875)
 -  added the full basissets to the orb checkpoint file (#863)
+-  return default for empty strings in option file (#873)
 
 Version 2021.2 and earlier
 ==========================